requirements.txt
customtkinter>=5.0.0
pystray>=0.19.0
Pillow>=10.0.0
httpx>=0.25.0
fastapi>=0.109.0
uvicorn>=0.27.0
